sample/back/get_home_expense_total/requirements.txt
sampleproject==2.0.0
google-cloud-firestore==2.0.2
python-dateutil==2.7.5